Move MailHandler enum to srcf.database
